Add release action
The release action is essentially a copy of the test action but with
smaller variations like the tag is not statically set to 'latest' but is
based on the tag that is pushed. The trigger is of course the pushing of
a tag rather than any branch.

The test action is changed to ignore tags.